inventoryupdated

+++ lib/models/Item.dart
@@ -2,7 +2,7 @@ class InventoryItem {
   final String name;
   final String category;
   final int quantity;
-  final int totalCost;
+  final double totalCost;
 
   InventoryItem( {required this.name,required this.totalCost, required this.category, required this.quantity, required double price});
 }
